# Running queries: Compress large columns before transfer