This is not necessarily a complete implementation, but it is fully operational as far as I've tested locally.

Curious to see what you think about this type of optimization.

I use DBADash primarily over a VPN connection. The Running Queries grid is regularly the slowest to load, especially now that I have plan collection enabled. After some testing, I found that the majority of the wait time was network IO.

By compressing the two larger columns (batch_text and text) and _only_ pulling the query plan when it's actually requested, that significantly improved usage over a VPN.

Some test case results from my local machine over a VPN connection:

| Running Query Count | Before (ms) | After (ms) | 
|---------------------|-------------|------------| 
| 5                   |  503        |  559       | 
| 10                  |  1,683      |  669       | 
| 20                  |  2,054      |  811       | 
| 50                  |  11,349     |  1,159     | 
| 100                 |  17,796     |  1,446     | 
| 200                 |  34,594     |  2,620     | 
| 400                 |  123,737    |  3,741     | 

Since my changes to `RunningQueries_Get` is not backward compatible, I could see changing it so that the compression of the columns are optionally set? This way if any users are relying on `RunningQueries_Get` to return the columns which were removed, they would stay put and a parameter could control whether they get swapped out with compressed versions?