@chadbaldwin 2024-12-27
I also had a bunch of other ideas going around in my head while building this to make it a more flexible change that could be used in many other places. But I didn't want to go too overboard with it and figured since this is a mostly behind the scenes change, it could always be changed later without breaking existing installations.

For example, you could have a standard naming convention coupled with the data type. So all columns in the repository database like "*_compressed varbinary(*)" could be detected and auto decompressed when pulled.

Or maybe there's a better way to approach it than having to build a bunch of compressed/decompressed logic into various views, funcs and procs, etc.