@DavidWiseman 2024-12-27
Thanks for the PR.  I would *guess* that most of the gains come from removing the query plan column - I might be wrong though.  It's already compressed so it makes sense to decompress in the app.  Also, the column is large and only needed when clicking to view the query plan - removing it completely as you have done is the best approach.  
I'm less certain that compressing batch_text and text will be universally beneficial as there is some overhead to compress/decompress - still it's probably worth it on slow connections and the extra overhead of compression is likely to be small in other cases.  I'm OK with doing this if you've tested the benefit.  Any extra work to compress/decompress the text is probably offset by the gains in not doing this for the query plan column.  

RunningQueries_Get doesn't need to be backward compatible.  As long as there is a major version bump, client apps will be prompted to update.

Options for further improvements (which don't need to be included in this PR)

The statement text is calculated from the batch_text which could also be done in the app, reducing the amount of data sent.  There might be an argument for storing the data compressed to reduce the storage cost.  There would be some overhead in compressing on insert, but this could be done in the app.  

I haven't tested any of the changes yet - these are just my initial thoughts.  It seems like a good fix, once you are happy with the PR, let me know and I'll give it a proper test when I get time.