@chadbaldwin 2024-12-27
Ah that's a good point. I'll play around with the code and see how much of an impact each of the changes on their own make.