@chadbaldwin 2024-12-30
I did a quick test a few days ago and I think you're right. Simply removing the query plan did make a pretty significant improvement on its own. There's still additional improvement by doing the pre-compression and decompression, but I would probably now argue that it's not enough to justify the change and introducing possible new issues.

I'll do a bit more testing, but I will more than likely strip this PR down to _just_ the query plan removal so that it's as simple a change as possible while making the largest impact.