@DavidWiseman 2025-04-29
Hi @chadbaldwin .  I'm just wondering if you have any more thoughts on this PR? If you don't have time to look at this, just let me know.  If not, I'll probably look to implement removing the query plan - just want to give you the opportunity to take credit for the work.  

I'm considering using the messaging feature to pull the query plan directly from the monitored instance where it's not available in the repository database.  It makes sense to remove the query plan column first or do it as part of this change.

The benefits of the text compression will vary but might still provide benefit over a slow VPN.  On one of my snapshots with 150 rows, compression reduced the text size from 7,578KB to 907KB, but added an extra ~150ms CPU time from SQL side (not including the extra decompression step on the client).  The text size would be 5,740KB, just sending the batch_text and calculating text on the client (or 677KB compressed).  Removing the query plan from this same snapshot saves 39,138KB with no real downsides.  

Storing the text data compressed is another option which would have the benefit of reducing the storage cost and removing the on the fly compression overhead.  It would add compression overhead to the collection though.  It will also add overhead to querying the text column with LIKE as we'd have to decompress it first (probably the main argument against doing it).  This would reduce the table size from 4.2GB to 0.8GB for one of my repos (a small % of the repo size).  

Thanks,

David