@chadbaldwin 2025-04-29
I'm personally less worried about storage and storage costs..._usually_ storage is cheap, so if anything I would rather store _more_ data that is pre-computed to make things faster, rather than finding ways to reduce it. But that's just me.

For example, I've considered building something that pre-computes the RunningQueries dataset. This way the UI is purely querying a single wide denormalized table, instead of reaching out to various tables, views and functions. Which could then be further indexed for the "View ALL" search feature.

Or another example would be storing large text data both as plain text as well as compressed if it meant speeding up data transfer between DB and client. If it means putting a little overhead on the collection data load side, that's better to do it once there than 10 times on the client side. But that could bulk the database up quite a bit.

As far as compressing text before sending over the wire...I did do some testing on this back when I started this work after your response. The benefits were not all that impressive. The majority of the performance gain was purely in lazy loading the plans on demand. There was some, but not much, additional benefit in compressing text before sending over the wire.

I think what I'll do with this PR is close it and then make a new one where I am ONLY making the plan lazy loading and caching change. I'll try to get this knocked out today since it's a fairly simple change.