@DavidWiseman 2025-04-30
Thank you.  A PR with just the plan lazy loading is easy to approve - it's 100% the right thing to do.  It will also reduce memory usage on the client.

For running queries, the capture process runs frequently on potentially hundreds instances.  It's important that the collection, import and storage is efficient.  It's also important that you can access the snapshots quickly.  Where we need to optimize one in the expense of the other, it's about finding a good balance.

RunningQueries_Get averages ~200ms for me and loading from the client side feels fast in most cases.  It doesn't make sense to optimize this further at the expense of extra load on collection/import or additional storage - based on these load times.  The low volume of query execution for RunningQueries_Get compared to the high volume of RunningQueries_Upd factors in to this.  Most snapshots are probably not loaded so the read to write factor is the other way around in this case.